+++ src/headers/row.h
@@ -4,13 +4,13 @@
 typedef struct
 {
 	char *guess;
-	short c_right_pegs;
-	short c_almost_right_pegs;
+	int c_right_pegs;
+	int c_almost_right_pegs;
 } row_t;
 
 void destroy_row(row_t *row);
 row_t *create_row();
 void set_guess(row_t *row, char *guess);
-void set_pebles(row_t *row, short c_right_pegs, short c_almost_right_pegs);
+void set_pebles(row_t *row, int c_right_pegs, int c_almost_right_pegs);
 
 #endif
